moved interface extensions to their own files
