+++ bitex/interface/poloniex.py
@@ -1,35 +1,90 @@
 # Import Built-Ins
 import logging
-import time
-
-# Import Third-Party
-import requests
 
 # Import Homebrew
-from bitex.exceptions import UnsupportedPairError
-
-from bitex.api.REST.bitfinex import BitfinexREST
-from bitex.api.REST.bitstamp import BitstampREST
-from bitex.api.REST.bittrex import BittrexREST
-from bitex.api.REST.bter import BterREST
-from bitex.api.REST.ccex import CCEXREST
-from bitex.api.REST.coincheck import CoincheckREST
-from bitex.api.REST.cryptopia import CryptopiaREST
-from bitex.api.REST.gdax import GDAXREST
-from bitex.api.REST.gemini import GeminiREST
-from bitex.api.REST.hitbtc import HitBTCREST
-from bitex.api.REST.itbit import ITbitREST
-from bitex.api.REST.kraken import KrakenREST
-from bitex.api.REST.okcoin import OKCoinREST
 from bitex.api.REST.poloniex import PoloniexREST
-from bitex.api.REST.quadriga import QuadrigaCXREST
-from bitex.api.REST.quoine import QuoineREST
-from bitex.api.REST.rocktrading import RockTradingREST
-from bitex.api.REST.vaultoro import VaultoroREST
-from bitex.api.REST.yunbi import YunbiREST
-
 from bitex.interface.rest import RESTInterface
-from bitex.utils import check_version_compatibility, check_and_format_pair
+from bitex.utils import check_and_format_pair
 
 # Init Logging Facilities
 log = logging.getLogger(__name__)
+
+
+class Poloniex(RESTInterface):
+    def __init__(self, **APIKwargs):
+        super(Poloniex, self).__init__('Poloniex', PoloniexREST(**APIKwargs))
+
+    def request(self, endpoint, authenticate=False, **req_kwargs):
+        if 'params' in req_kwargs:
+            req_kwargs['params'].update({'command': endpoint})
+        else:
+            req_kwargs['params'] = {'command': endpoint}
+        if authenticate:
+            return super(Poloniex, self).request('POST', endpoint, authenticate,
+                                                 **req_kwargs)
+        else:
+            return super(Poloniex, self).request('GET', 'public', authenticate,
+                                                 **req_kwargs)
+
+    def _get_supported_pairs(self):
+        return ['BTC_ETH']
+
+    # Public Endpoints
+    @check_and_format_pair
+    def ticker(self, pair, *args, **kwargs):
+        return self.request('returnTicker', params=kwargs)
+
+    @check_and_format_pair
+    def order_book(self, pair, *args, **kwargs):
+        payload = {'currencyPair': pair}
+        payload.update(kwargs)
+        return self.request('returnOrderBook', params=payload)
+
+    @check_and_format_pair
+    def trades(self, pair, *args, **kwargs):
+        payload = {'currencyPair': pair}
+        payload.update(kwargs)
+        return self.request('returnTradeHistory', params=payload)
+
+    # Private Endpoints
+    def _place_order(self, pair, price, size, side, **kwargs):
+        payload = {'currencyPair': pair, 'rate': price, 'amount': size}
+        payload.update(kwargs)
+        if side == 'bid':
+            return self.request('buy', authenticate=True, params=payload)
+        else:
+            return self.request('sell', authenticate=True, params=payload)
+
+    @check_and_format_pair
+    def ask(self, pair, price, size, *args, **kwargs):
+        raise NotImplementedError
+
+    @check_and_format_pair
+    def bid(self, pair, price, size, *args, **kwargs):
+        raise NotImplementedError
+
+    def order_status(self, order_id, *args, **kwargs):
+        payload = {'orderNumber': order_id}
+        payload.update(kwargs)
+        return self.request('returnOrderTrades', authenticate=True,
+                            params=payload)
+
+    def open_orders(self, *args, **kwargs):
+        payload = {'currencyPair': 'all'}
+        payload.update(kwargs)
+        return self.request('returnOpenOrders', authenticate=True,
+                            params=payload)
+
+    def cancel_order(self, *order_ids, **kwargs):
+        results = []
+        payload = kwargs
+        for oid in order_ids:
+            payload.update({'orderNumber', oid})
+            r = self.request('cancelOrder', authenticate=True, params=oid)
+            results.append(r)
+        return results if len(results) > 1 else results[0]
+
+    def wallet(self, *args, **kwargs):
+        return self.request('returnTradableBalances', authenticate=True,
+                            params=kwargs)
+
