+++ bitex/interface/rocktrading.py
@@ -1,35 +1,102 @@
 # Import Built-Ins
 import logging
-import time
-
-# Import Third-Party
-import requests
 
 # Import Homebrew
-from bitex.exceptions import UnsupportedPairError
-
-from bitex.api.REST.bitfinex import BitfinexREST
-from bitex.api.REST.bitstamp import BitstampREST
-from bitex.api.REST.bittrex import BittrexREST
-from bitex.api.REST.bter import BterREST
-from bitex.api.REST.ccex import CCEXREST
-from bitex.api.REST.coincheck import CoincheckREST
-from bitex.api.REST.cryptopia import CryptopiaREST
-from bitex.api.REST.gdax import GDAXREST
-from bitex.api.REST.gemini import GeminiREST
-from bitex.api.REST.hitbtc import HitBTCREST
-from bitex.api.REST.itbit import ITbitREST
-from bitex.api.REST.kraken import KrakenREST
-from bitex.api.REST.okcoin import OKCoinREST
-from bitex.api.REST.poloniex import PoloniexREST
-from bitex.api.REST.quadriga import QuadrigaCXREST
-from bitex.api.REST.quoine import QuoineREST
 from bitex.api.REST.rocktrading import RockTradingREST
-from bitex.api.REST.vaultoro import VaultoroREST
-from bitex.api.REST.yunbi import YunbiREST
-
 from bitex.interface.rest import RESTInterface
-from bitex.utils import check_version_compatibility, check_and_format_pair
+from bitex.utils import check_and_format_pair
 
 # Init Logging Facilities
 log = logging.getLogger(__name__)
+
+
+class TheRockTrading(RESTInterface):
+    def __init__(self, **APIKwargs):
+        super(TheRockTrading, self).__init__('The Rock Trading Ltd.',
+                                             RockTradingREST(**APIKwargs))
+
+    def _get_supported_pairs(self):
+        return [d['id'] for d in self.request('GET', 'funds').json()['funds']]
+
+    # Public Endpoints
+    @check_and_format_pair
+    def ticker(self, pair, *args, **kwargs):
+        return self.request('GET', 'funds/%s/ticker' % pair, params=kwargs)
+
+    @check_and_format_pair
+    def order_book(self, pair, *args, **kwargs):
+        return self.request('GET', 'funds/%s/orderbook' % pair, params=kwargs)
+
+    @check_and_format_pair
+    def trades(self, pair, *args, **kwargs):
+        return self.request('GET', 'funds/%s/trades' % pair, params=kwargs)
+
+    # Private Endpoints
+    def _place_order(self, pair, price, size, side, **kwargs):
+        payload = {'price': price, 'amount': size, 'side': side}
+        payload.update(kwargs)
+        return self.request('POST', 'funds/%s/orders' % pair, authenticate=True,
+                            params=payload)
+
+    @check_and_format_pair
+    def ask(self, pair, price, size, *args, **kwargs):
+        return self._place_order(pair, price, size, 'sell', **kwargs)
+
+    @check_and_format_pair
+    def bid(self, pair, price, size, *args, **kwargs):
+        return self._place_order(pair, price, size, 'buy', **kwargs)
+
+    def order_status(self, order_id, *args, **kwargs):
+        if 'pair' not in kwargs and 'fund_id' not in kwargs:
+            raise ValueError("Need to specify pair or fund_id in kwargs!")
+        else:
+            try:
+                pair = kwargs.pop('pair')
+            except KeyError:
+                pair = kwargs.pop('fund_id')
+            return self.request('GET', 'funds/%s/orders/%s' % (pair, order_id),
+                                authenticate=True, params=kwargs)
+
+    def open_orders(self, *args, **kwargs):
+        if 'pair' not in kwargs and 'fund_id' not in kwargs:
+            results = []
+            for fund_id in self.supported_pairs:
+                r = self.request('GET', 'funds/%s/orders' % fund_id,
+                                 authenticate=True, params=kwargs)
+                results.append(r)
+            return results if len(results) > 1 else results[0]
+        else:
+            try:
+                pair = kwargs.pop('pair')
+            except KeyError:
+                pair = kwargs.pop('fund_id')
+
+            return self.request('GET', 'funds/%s/orders' % pair,
+                                authenticate=True, params=kwargs)
+
+    def cancel_order(self, *order_ids, all=False, **kwargs):
+        results = []
+        if 'pair' not in kwargs and 'fund_id' not in kwargs:
+            for fund_id in self.supported_pairs:
+                r = self.cancel_order(*order_ids, all, pair=fund_id)
+                results.append(r)
+            return results if len(results) > 1 else results[0]
+        else:
+            try:
+                pair = kwargs.pop('pair')
+            except KeyboardInterrupt:
+                pair = kwargs.pop('fund_id')
+
+        if all:
+            return self.request('DELETE', 'funds/%s/orders/remove_all' % pair,
+                                authenticate=True, params=kwargs)
+        else:
+            for oid in order_ids:
+                r = self.request('DELETE', 'funds/%s/orders/%s' % (pair, oid),
+                                 authenticate=True, params=kwargs)
+                results.append(r)
+            return results if len(results) > 1 else results[0]
+
+    def wallet(self, *args, **kwargs):
+        return self.request('GET', 'balances', authenticate=True, params=kwargs)
+
